+++ Makefile
@@ -1,11 +1,15 @@
 .PHONY: client server
 
-all: client server
+all: common client server 
 	mv client/simpleftp-cl .
 	mv server/simpleftp-sv .
 	
+common:
+	+$(MAKE) -C common
+	
 client: 
 	+$(MAKE) -C client
+
 server:
 	+$(MAKE) -C server
