+++ client/Makefile
@@ -1,4 +1,4 @@
 CC=gcc
 CFLAGS= -std=gnu99 -Wall
 all: simpleftp-cl
-simpleftp-cl: helper-cl.c simpleftp-cl.h simpleftp-cl.c
+simpleftp-cl: helper-cl.c simpleftp-cl.h simpleftp-cl.c ../common/common.c
