+++ client/helper-cl.c
@@ -1,15 +1,5 @@
 #include "simpleftp-cl.h"
 
-/* make_socket
- * Create a socket
- * returns: socket descrriptor
- */
-int make_socket(void) {
-	int sock;
-	sock = socket(PF_INET, SOCK_STREAM, 0);
-	if (sock < 0) ERR("socket");
-	return sock;
-}
 
 /* make_address
  * Resolve the address given the domain name(name) and the port.
@@ -38,7 +28,7 @@ struct sockaddr_in make_address(char *name, char *port) {
 int connect_socket(char *name, char *port) {
 	struct sockaddr_in addr;
 	int socketfd;
-	socketfd = make_socket();
+	socketfd = make_socket(PF_INET, SOCK_STREAM);
 	addr = make_address(name, port);
 	if (connect(socketfd, (struct sockaddr *) &addr, sizeof(struct sockaddr_in)) < 0) {
 		if (errno != EINTR) {
@@ -62,53 +52,4 @@ int connect_socket(char *name, char *port) {
 		}
 	}
 	return socketfd;
-}
-
-
-/* set_handler
- * Set a function as a handler for signal
- * void (*f)(int) - pointer to function
- * int sigNo - symbolic constant of a signal
- * return: status
- */
-int set_handler( void (*f)(int), int sigNo) {
-	struct sigaction act;
-	memset(&act, 0, sizeof(struct sigaction));
-	act.sa_handler = f;
-	if (-1 == sigaction(sigNo, &act, NULL)) {
-		return -1;
-	}
-	return 0;
-}
-
-/* persist_read
- * Read up to |count| bytes from |fd| into |buf|.
- * Persist if interrupted by a signal.
- */
-ssize_t persist_read(int fd, char *buf, size_t count) {
-	ssize_t status;
-	if ( (status = TEMP_FAILURE_RETRY(read(fd, buf, count))) < 0) {
-		ERR("read");
-	}
-	return status;
-}
-
-/* persist_write
- * Write |count| bytes from |buf| to |fd|.
- * Persist if interrupted by a signal.
- */
-ssize_t persist_write(int fd, char *buf, size_t count) {
-	int c;
-	size_t len = 0;
-	do {
-		c = TEMP_FAILURE_RETRY(write(fd, buf, count));
-		if(c < 0) {
-			ERR("write");
-		}
-		buf += c;
-		len += c;
-		count -= c;
-	}
-	while (count > 0);
-	return len;
 }
