+++ server/Makefile
@@ -2,4 +2,4 @@ CC=gcc
 CFLAGS= -std=gnu99 -Wall
 LDLIBS= -lpthread
 all: simpleftp-sv
-simpleftp-sv: helper-sv.c thread-sv.c simpleftp-sv.h simpleftp-sv.c
+simpleftp-sv: helper-sv.c thread-sv.c simpleftp-sv.h simpleftp-sv.c ../common/common.c
